edicao codigo singleton
