+++ code/singleton/.env
@@ -1,0 +1,11 @@
+# database service config
+POSTGRES_PASSWORD="pass"
+POSTGRES_USER="user"
+POSTGRES_DB="singleton_db"
+
+# connector config
+DB_HOST="db"
+DB_PORT=5432
+DB_USER="user"
+DB_PASS="pass"
+DB_BASE="singleton_db"
