+++ code/singleton/README.md
@@ -1,0 +1,17 @@
+.env example:
+
+```bash
+
+# database service config
+POSTGRES_PASSWORD="zeroo_sqs"
+POSTGRES_USER="zeroo_sqs"
+POSTGRES_DB="zeroo_sqs"
+
+# api/migrator service config
+DB_HOST="db"
+DB_PORT="5432"
+DB_USER="zeroo_sqs"
+DB_PASS="zeroo_sqs"
+DB_BASE="zeroo_sqs"
+
+```
