+++ code/singleton/requirements.txt
@@ -1,0 +1,2 @@
+psycopg2-binary==2.9.9
+SQLAlchemy==2.0.35
